+++ packages/react-loader/CHANGELOG.md
@@ -30,6 +30,16 @@ All notable changes to this project will be documented in this file. See
   * dependencies
     * @sanity/core-loader bumped from 1.6.0 to 1.6.1
 
+## [1.11.2](https://github.com/sanity-io/visual-editing/compare/react-loader-v1.11.1...react-loader-v1.11.2) (2025-04-10)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @sanity/core-loader bumped to 1.8.2
+    * @sanity/visual-editing-csm bumped to 2.0.12
+
 ## [1.11.1](https://github.com/sanity-io/visual-editing/compare/react-loader-v1.11.0...react-loader-v1.11.1) (2025-04-01)